Use @staticmethod, add mutually exclusive flags
Instead of == None, it is more idiomatic to use is None. 

If you specify the number of args for argparse, argparse will automatically create a list of the specified size - there is no need to check the length of the output. You can also use parser.add_mutually_exclusive_group() to create a mutually exclusive group, of which only one option can be specified, which helps simplify your Arguments class a lot.

If you use a number of methods together, such as getting the title, page, and format of a given movie, it may be better to just put them into a single function - the user probably doesn't care about the page, which is raw data, but is interested in knowing the title and available formats of their movie.

For functions that should belong inside a class but don't make use of self, it's often a good idea to use @staticmethod to enforce this - it helps avoid potential bugs from modifying the class, and it allows the method to be used as a class level method - eg. YTS.static_method(arg1, arg2), without having to instantiate a YTS object.

I also made use of __getitem__ for Arguments, which defines how Arguments["a"] behaves - so we can now directly call args["p"] without having to specify .arguments, which looks strange if it's already an Arguments class. Alternatively, by using @staticmethod, we could also call Arguments.get_cli_arguments() and avoid creating an Argument object altogether.